OTHER REFACTOR Extracted Login function into LoginViewController class
