+++ messenger-app/UserViewController.swift
@@ -15,22 +15,6 @@ class UserViewController: UIViewController {
     @IBOutlet weak var usernameText: UITextField!
     @IBOutlet weak var emailText: UITextField!
     @IBOutlet weak var passwordText: UITextField!
-
-    @IBAction func logIn(_ sender: UIButton) {
-        let email = emailText.text!
-        let password = passwordText.text!
-        database.signIn(withEmail: email, password: password, completion: { ( user, error) in
-            if user != nil {
-                self.performSegue(withIdentifier: "Log in successful", sender: self)
-            }
-            else {
-                let alertMessage = error?.localizedDescription ?? "Please enter a valid email address an password"
-                let alert = UIAlertController(title: "Failed to log in", message: alertMessage, preferredStyle: .alert)
-                alert.addAction(UIAlertAction(title: "OK", style: .cancel, handler: nil))
-                self.present(alert, animated: true)
-            }
-        })
-    }
     
     @IBAction func signUp(_ sender: UIButton) {
         let username = usernameText.text!
